Fix (another) overflow when seeking in long WMA files.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@16597 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/wma.c
@@ -403,7 +403,7 @@ static int seek(int ms, asf_waveformatex_t* wfx)
 
     /*estimate packet number from bitrate*/
     int initial_packet = ci->curpos/wfx->packet_size;
-    int packet_num = (ms*(wfx->bitrate>>3))/wfx->packet_size/1000;
+    int packet_num = (((int64_t)ms)*(wfx->bitrate>>3))/wfx->packet_size/1000;
     int last_packet = ci->id3->filesize / wfx->packet_size;
 
     if (packet_num > last_packet) {
